modularizing ColumnEditMenu
